Tidy up browserstack and local files into separate directories to clear out the top level directory.
