+++ src/test/java/uk/gov/dhsc/htbhf/CucumberConfiguration.java
@@ -3,6 +3,8 @@ package uk.gov.dhsc.htbhf;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.PropertySource;
+import uk.gov.dhsc.htbhf.browserstack.BrowserStackConfiguration;
+import uk.gov.dhsc.htbhf.local.LocalConfiguration;
 
 @Configuration
 @PropertySource("classpath:application.properties")
